api/requirements.txt
coverage~=7.2.4
beautifulsoup4==4.12.2
flask~=2.3.2
Flask-SQLAlchemy~=3.0.3
flask-login==0.6.2
flask-migrate~=4.0.4
flask-restful==0.3.9
flask-session2==1.3.1
flask-cors==3.0.10
gunicorn~=20.1.0
gevent~=22.10.2
langchain==0.0.142
llama-index==0.5.27
openai~=0.27.5
psycopg2-binary~=2.9.6
pycryptodome==3.17
python-dotenv==1.0.0
pytest~=7.3.1
tiktoken==0.3.3
Authlib==1.2.0
boto3~=1.26.123
tenacity==8.2.2
cachetools~=5.3.0
weaviate-client~=3.16.2
qdrant_client~=1.1.6
pinecone-client~=2.2.1
mailchimp-transactional~=1.0.50
scikit-learn==1.2.2
sentry-sdk[flask]~=1.21.1
jieba==0.42.1
celery==5.2.7
redis~=4.5.4
pypdf==3.8.1